Add syntax for absorbing free vars in a copy of scope
