+++ legacy/src/main/scala/coop/rchain/rholang/normalizer2/syntax/EffectSyntax.scala
@@ -1,6 +1,9 @@
 package coop.rchain.rholang.normalizer2.syntax
 
-import coop.rchain.rholang.normalizer2.env.{BoundVarScope, FreeVarScope}
+import cats.effect.Sync
+import cats.implicits.{toFlatMapOps, toFunctorOps}
+import coop.rchain.rholang.interpreter.compiler.FreeContext
+import coop.rchain.rholang.normalizer2.env.{BoundVarScope, BoundVarWriter, FreeVarScope}
 
 trait EffectSyntax {
   implicit def normalizerEffectSyntax[F[_], A](f: F[A]): NormalizerEffectOps[F, A] = new NormalizerEffectOps[F, A](f)
@@ -17,4 +20,30 @@ class NormalizerEffectOps[F[_], A](val f: F[A]) extends AnyVal {
     insideReceive: Boolean = false,
   )(implicit fwScope: FreeVarScope[F], bwScope: BoundVarScope[F]): F[A] =
     bwScope.withNewBoundVarScope(fwScope.withNewFreeVarScope(insideReceive)(f))
+
+  /** Bound free variables in a copy of the current scope.
+   *
+   * Free variables are sorted by levels and then added with indexes:
+   * {i0, i1, ..., iN} = {fl0 + last + 1, fl1 + last + 1, ..., flN + last + 1}.
+   * Here, i0, ..., iN represent the Bruijn indices of the new bound vars,
+   * fl0, ..., flN are the Bruijn levels of the inserted free vars,
+   * last is the last index among all bound vars at the moment. */
+
+  def withAbsorbedFreeVars[T](
+    freeVars: Seq[(String, FreeContext[T])],
+  )(implicit sync: Sync[F], bwScope: BoundVarScope[F], bwWriter: BoundVarWriter[T]): F[A] = {
+
+    def absorbFree(freeVars: Seq[(String, FreeContext[T])]): Seq[Int] = {
+      val sortedByLevel  = freeVars.sortBy(_._2.level)
+      val (levels, data) = sortedByLevel.unzip(fv => (fv._2.level, (fv._1, fv._2.typ, fv._2.sourcePosition)))
+      assert(
+        levels == levels.indices,
+        "Error when absorbing free variables during normalization: incorrect de Bruijn levels." +
+          s"Should be ${levels.indices}, but was $levels.",
+      )
+      bwWriter.putBoundVars(data)
+    }
+
+    bwScope.withCopyBoundVarScope(sync.delay(absorbFree(freeVars)).flatMap(_ => f))
+  }
 }
