+++ legacy/src/main/scala/coop/rchain/rholang/package.scala
@@ -2,7 +2,7 @@ package coop.rchain
 
 import coop.rchain.metrics.Metrics
 import coop.rchain.rholang.interpreter.{RhoHistoryRepositorySyntax, RhoRuntimeSyntax}
-import coop.rchain.rholang.normalizer2.syntax.{BoundVarWriterSyntax, EffectSyntax}
+import coop.rchain.rholang.normalizer2.syntax.EffectSyntax
 
 package object rholang {
   val RholangMetricsSource: Metrics.Source = Metrics.Source(Metrics.BaseSource, "rholang")
@@ -10,8 +10,4 @@ package object rholang {
   object syntax extends AllSyntaxRholang
 }
 
-trait AllSyntaxRholang
-    extends RhoRuntimeSyntax
-    with RhoHistoryRepositorySyntax
-    with BoundVarWriterSyntax
-    with EffectSyntax
+trait AllSyntaxRholang extends RhoRuntimeSyntax with RhoHistoryRepositorySyntax with EffectSyntax
